+++ R/batchtools_template.R
@@ -49,7 +49,11 @@ class(batchtools_lsf) <- c(
   "batchtools_multiprocess", "batchtools",
   "multiprocess", "future", "function"
 )
-attr(batchtools_lsf, "tweakable") <- c("finalize")
+attr(batchtools_lsf, "tweakable") <- c(
+  "finalize",
+  ## Arguments to batchtools::makeClusterFunctionsLSF()
+  "scheduler.latency", "fs.latency"
+)
 
 #' @export
 #' @rdname batchtools_template
@@ -70,7 +74,11 @@ class(batchtools_openlava) <- c(
   "batchtools_multiprocess", "batchtools",
   "multiprocess", "future", "function"
 )
-attr(batchtools_openlava, "tweakable") <- c("finalize")
+attr(batchtools_openlava, "tweakable") <- c(
+  "finalize",
+  ## Arguments to batchtools::makeClusterFunctionsOpenLava()
+  "scheduler.latency", "fs.latency"
+)
 
 #' @export
 #' @rdname batchtools_template
@@ -91,7 +99,11 @@ class(batchtools_sge) <- c(
   "batchtools_multiprocess", "batchtools",
   "multiprocess", "future", "function"
 )
-attr(batchtools_sge, "tweakable") <- c("finalize")
+attr(batchtools_sge, "tweakable") <- c(
+  "finalize",
+  ## Arguments to batchtools::makeClusterFunctionsSGE()
+  "nodename", "scheduler.latency", "fs.latency"
+)
 
 #' @export
 #' @rdname batchtools_template
@@ -112,7 +124,12 @@ class(batchtools_slurm) <- c(
   "batchtools_multiprocess", "batchtools",
   "multiprocess", "future", "function"
 )
-attr(batchtools_slurm, "tweakable") <- c("finalize")
+attr(batchtools_slurm, "tweakable") <- c(
+  "finalize",
+  ## Arguments to batchtools::makeClusterFunctionsSlurm()
+  "array.jobs", "nodename", "scheduler.latency", "fs.latency"
+)
+
 
 #' @export
 #' @rdname batchtools_template
@@ -133,7 +150,11 @@ class(batchtools_torque) <- c(
   "batchtools_multiprocess", "batchtools",
   "multiprocess", "future", "function"
 )
-attr(batchtools_torque, "tweakable") <- c("finalize")
+attr(batchtools_torque, "tweakable") <- c(
+  "finalize",
+  ## Arguments to batchtools::makeClusterFunctionsTORQUE()
+  "scheduler.latency", "fs.latency"
+)
 
 
 #' @importFrom batchtools makeClusterFunctionsLSF
@@ -152,6 +173,8 @@ batchtools_by_template <- function(expr, envir = parent.frame(),
   if (substitute) expr <- substitute(expr)
   type <- match.arg(type)
 
+  dotdotdot <- list(...)
+
   make_cfs <- switch(type,
     lsf      = makeClusterFunctionsLSF,
     openlava = makeClusterFunctionsOpenLava,
@@ -160,28 +183,37 @@ batchtools_by_template <- function(expr, envir = parent.frame(),
     torque   = makeClusterFunctionsTORQUE
   )
 
+  make_cfs_formals <- formals(make_cfs)
+  
   ## Get the default template?
   if (is.null(template)) {
-    template <- formals(make_cfs)$template
+    template <- make_cfs_formals$template
   }
 
   stop_if_not(is.character(template), length(template) == 1L,
               !is.na(template), nzchar(template))
 
   template <- find_template_file(template)
-  cluster.functions <- make_cfs(template)
+
+  keep <- which(names(dotdotdot) %in% names(make_cfs_formals))
+  args <- c(list(template = template), dotdotdot[keep])
+  cluster.functions <- do.call(make_cfs, args = args)
   attr(cluster.functions, "template") <- template
 
-  future <- BatchtoolsTemplateFuture(
+  ## Drop used '...' arguments
+  if (length(keep) > 0) dotdotdot <- dotdotdot[-keep]
+
+  args <- list(
     expr = expr, envir = envir, substitute = FALSE,
     globals = globals,
     label = label,
     cluster.functions = cluster.functions,
     registry = registry,
     resources = resources,
-    workers = workers,
-    ...
+    workers = workers
   )
+  if (length(dotdotdot) > 0) args <- c(args, dotdotdot)
+  future <- do.call(BatchtoolsTemplateFuture, args = args)
 
   if (!future$lazy) future <- run(future)